+++ src/App.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import { Route, Routes } from "react-router-dom";
 import Layout from "./components/layout/Layout";
 import { ArtProvider } from "./context/ArtContext";
-import Detail from "./pages/Detail";
-import Favorites from "./pages/Favorites";
+import Detail from "./pages/Detail/Detail";
+import Favorites from "./pages/Favorites/Favorites";
 import Home from "./pages/Home";
 
 const App: React.FC = () => {
