+++ src/components/layout/Layout.tsx
@@ -3,8 +3,18 @@ import styled from "styled-components";
 import GlobalStyles from "../../styles/GlobalStyles";
 import ErrorBoundary from "../shared/ErrorBoundary";
 import Container from "./Container";
-import Footer from "./Footer";
-import Header from "./Header";
+import Footer from "./Footer/Footer";
+import Header from "./Header/Header";
+
+const PageWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+`;
+
+const ContentWrapper = styled.main`
+  flex: 1;
+`;
 
 const Layout: React.FC = () => {
   return (
@@ -24,13 +34,3 @@ const Layout: React.FC = () => {
 };
 
 export default Layout;
-
-const PageWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-`;
-
-const ContentWrapper = styled.main`
-  flex: 1;
-`;
