+++ src/components/sections/OtherWorks.tsx
@@ -1,9 +1,20 @@
 import React, { useContext } from "react";
 import styled from "styled-components";
 import ArtContext from "../../context/ArtContext";
-import SmallCard from "../ui/SmallCard";
+import SmallCard from "../ui/SmallCard/SmallCard";
 import Spinner from "../ui/Spinner";
 import TitleSection from "../ui/TitleSection";
+
+const Wrapper = styled.div`
+  padding: 20px 0 100px 0;
+  text-align: center;
+`;
+
+const Grid = styled.div`
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(340px, 1fr));
+  gap: 20px;
+`;
 
 const OtherWorks: React.FC = () => {
   const { artworks, loading, error } = useContext(ArtContext);
@@ -35,14 +46,3 @@ const OtherWorks: React.FC = () => {
 };
 
 export default OtherWorks;
-
-const Wrapper = styled.div`
-  padding: 20px 0 100px 0;
-  text-align: center;
-`;
-
-const Grid = styled.div`
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(340px, 1fr));
-  gap: 20px;
-`;
