+++ src/components/ui/Pagination.tsx
@@ -3,6 +3,38 @@ import styled from "styled-components";
 import { INITIAL_PAGE_INDEX, MAX_VISIBLE_PAGES } from "../../constants";
 import { PaginationProps } from "../../types";
 
+const PaginationWrapper = styled.div`
+  display: flex;
+  justify-content: flex-end;
+  width: 100%;
+  margin: 20px 0;
+`;
+
+const PageButton = styled.button`
+  background: #fafafa;
+  color: #333;
+  border: none;
+  border-radius: 5px;
+  padding: 5px 10px;
+  margin: 0 5px;
+  cursor: pointer;
+
+  &:hover {
+    background: #ff7a00;
+    color: #fff;
+  }
+
+  &.active {
+    background: #ff7a00;
+    color: #fff;
+  }
+
+  &:disabled {
+    opacity: 0;
+    cursor: default;
+  }
+`;
+
 const Pagination: React.FC<PaginationProps> = ({
   currentPage,
   totalPages,
@@ -67,35 +99,3 @@ const Pagination: React.FC<PaginationProps> = ({
 };
 
 export default Pagination;
-
-const PaginationWrapper = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  width: 100%;
-  margin: 20px 0;
-`;
-
-const PageButton = styled.button`
-  background: #fafafa;
-  color: #333;
-  border: none;
-  border-radius: 5px;
-  padding: 5px 10px;
-  margin: 0 5px;
-  cursor: pointer;
-
-  &:hover {
-    background: #ff7a00;
-    color: #fff;
-  }
-
-  &.active {
-    background: #ff7a00;
-    color: #fff;
-  }
-
-  &:disabled {
-    opacity: 0;
-    cursor: default;
-  }
-`;
