+++ src/pages/Home.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import OtherWorks from "../components/sections/OtherWorks";
 import SearchForm from "../components/sections/SearchForm/SearchForm";
-import Topics from "../components/sections/Topics";
+import Topics from "../components/sections/Topics/Topics";
 
 const Home: React.FC = () => {
   return (
